+++ system/src/Grav/Common/Config/Setup.php
@@ -401,7 +401,7 @@ class Setup extends Data
             }
 
             // Create security.yaml salt if it doesn't exist into existing configuration environment if possible.
-            $securityFile = basename(static::$securityFile);
+            $securityFile = Utils::basename(static::$securityFile);
             $securityFolder = substr(static::$securityFile, 0, -\strlen($securityFile));
             $securityFolder = $locator->findResource($securityFolder, true) ?: $locator->findResource($securityFolder, true, true);
             $filename = "{$securityFolder}/{$securityFile}";