+++ system/src/Grav/Common/Media/Traits/MediaUploadTrait.php
@@ -132,9 +132,9 @@ trait MediaUploadTrait
             if ($folder === '.') {
                 $folder = '';
             }
-            $filename = basename($filename);
+            $filename = Utils::basename($filename);
         }
-        $extension = pathinfo($filename, PATHINFO_EXTENSION);
+        $extension = Utils::pathinfo($filename, PATHINFO_EXTENSION);
 
         // Decide which filename to use.
         if ($settings['random_name']) {